chore: move application queries to a class
